Updating the dropdown label substring handling
I moved the label logic into its own method.  I updated it to only substring and append the ellipsis if the label text is greater than 10 characters.  It also does this to the default label when none of the dropdown items are selected (so if the custom default label is too long, it trims it appropriately).