Fixing dropdown selection issue when the screen is resized
I migrated the rendering of the dropdown label into its own method so I could re-use it.

I added a method that is called on window resize to verify the selected tab is within the dropdown.  This way, if I select a tab and shrink the window size until it is in the dropdown, the dropdown updates to reflect that.  Likewise, if I select a dropdown item and resize the window so the item is now a visible tab, the dropdown updates to reflect that as well.