Added option to allow the dropdown tab to not require pull-right
I added a "pullDropdownRight" option that defaults to true (to maintain previous behavior by default).  When enabled, the plugin will add a "pull-right" class to the dropdown (so it doesn't need to be manually added to the tab).  When disabled, the plugin will move the dropdown tab after bringing the visible tabs back out so it remains at the end of the tab list.  This keeps the dropdown tab next to the other tabs instead of potentially having a gap between tabs.

I also added a call to checkDropdownSelection to the setup of the tabs so if a dropdown item is selected on-init, the dropdown tab correctly reflects this.